Create Code_EEMD_LSTM_LR.py
This code is to use ensemble EMD to decompose the original signal into several subsequences. Then, LSTM is performed on each subsequence to make predictions. Linear regression is used to get the weights of each subsequence taking up in the original signal. Finally, the fusion technique will fuse the predictions together to the final predictions.